# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Mote Agent Runtime API
  description: |
    Mote 是一个本地优先的 AI 代理运行时，提供 HTTP API 用于对话、会话管理、工具调用、记忆存储等功能。
    
    ## 认证
    API 使用 Bearer Token 认证，Token 从配置的 Copilot Provider 获取。
    
    ## 版本控制
    所有 API 端点以 `/api/v1` 为前缀，支持版本协商。
  version: "1.0.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Mote Team
    url: https://github.com/mote-agent/mote
servers:
  - url: http://localhost:18788/api/v1
    description: 本地开发服务器
  - url: https://mote.local/api/v1
    description: 局域网服务器
...
paths:
...
  /cron/jobs:
    get:
      tags: [cron]
      summary: 列出定时任务
      operationId: listCronJobs
      responses:
        "200":
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CronJobListResponse"
    post:
      tags: [cron]
      summary: 创建定时任务
      operationId: createCronJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCronJobRequest"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CronJob"
  /cron/jobs/{name}:
    get:
      tags: [cron]
      summary: 获取定时任务详情
      operationId: getCronJob
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CronJob"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [cron]
      summary: 更新定时任务
      operationId: updateCronJob
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCronJobRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CronJob"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [cron]
      summary: 删除定时任务
      operationId: deleteCronJob
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 删除成功
        "404":
          $ref: "#/components/responses/NotFound"
  /cron/jobs/{name}/run:
    post:
      tags: [cron]
      summary: 立即执行定时任务
      operationId: runCronJob
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 执行结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CronJobRunResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
    NotFound:
      description: 资源未找到
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    CronJob:
      type: object
      required: [name, schedule, type]
      properties:
        name:
          type: string
        schedule:
          type: string
          description: Cron 表达式
        type:
          type: string
          enum: [chat, tool, webhook]
        payload:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        last_run:
          type: string
          format: date-time
        next_run:
          type: string
          format: date-time
    CronJobListResponse:
      type: object
      required: [jobs]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/CronJob"
    CreateCronJobRequest:
      type: object
      required: [name, schedule, type]
      properties:
        name:
          type: string
        schedule:
          type: string
        type:
          type: string
          enum: [chat, tool, webhook]
        payload:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
          default: true
    UpdateCronJobRequest:
      type: object
      properties:
        schedule:
          type: string
        type:
          type: string
          enum: [chat, tool, webhook]
        payload:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
    CronJobRunResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        result:
          type: string
        error:
          type: string
        duration_ms:
          type: integer
...